pyproject.toml
[project]
name = "langgraph-coding-agent-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "e2b-code-interpreter>=2.0.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.0",
    "python-dotenv>=1.1.1",
]
